feat(create): implement AI Meme Generator component
- Add CreateClient component for uploading images and generating memes using AI
- Integrate image upload functionality with drag-and-drop support
- Implement meme generation logic with canvas rendering and caption handling
- Update CreatePage to utilize CreateClient component for a cleaner structure
